requirements.txt
beautifulsoup4==4.12.2
playwright==1.34.0
pytesseract==0.3.10
python-dotenv==1.0.0